+++ skrub/_reporting/tests/test_table_report.py
@@ -19,8 +19,8 @@ def test_report(air_quality):
     report = TableReport(air_quality, title="the title", column_filters=col_filt)
     html = report.html()
     assert "the title" in html
-    assert "Columns with null values" in html
-    assert "First 10 columns" in html
+    assert "With nulls" in html
+    assert "First 10" in html
     assert "First 2" in html
     for col_name in sbd.column_names(air_quality):
         assert col_name in html
@@ -74,3 +74,10 @@ def test_empty_dataframe(df_module):
 def test_open(pd_module, browser_mock):
     TableReport(pd_module.example_dataframe, title="the title").open()
     assert b"the title" in browser_mock.content
+
+
+def test_non_hashable_values(df_module):
+    # non-regression test for #1066
+    df = df_module.make_dataframe(dict(a=[[1, 2, 3], None, [4]]))
+    html = TableReport(df).html()
+    assert "[1, 2, 3]" in html