+++ CHANGES.rst
@@ -14,11 +14,50 @@ It is currently undergoing fast development and backward compatibility is not en
 
 Major changes
 -------------
+<<<<<<< HEAD
 * The :class:`TextEncoder` is now available to encode columns with diverse entries.
   It allows the representation of table entries as embeddings computed by a deep
   learning model. The weights of this model can be fetched locally or from the
   HuggingFace Hub.
   :pr:`1077` by :user:`Vincent Maladiere <Vincent-Maladiere>`.
+=======
+
+Minor changes
+-------------
+
+* The column filter selection dropdown in the tablereport is smaller and its
+  label has been removed to save space. :pr:`1107` by :user:`Jérôme Dockès
+  <jeromedockes>`.
+
+* The TableReport now uses the font size of its parent element when inserted
+  into another page. This makes it smaller in pages that use a smaller font size
+  than the browser default such as VSCode in some configurations. It also makes
+  it easier to control its size when inserting it in a web page by setting the
+  font size of its parent element. A few other small adjustments have also been
+  made to make it a bit more compact. :pr:`1098` by :user:`Jérôme Dockès
+  <jeromedockes>`.
+
+* Display of labels in the plots of the TableReport, especially for other
+  scripts than the latin alphabet, has improved.
+  - before, some characters could be missing and replaced by empty boxes.
+  - before, when the text is truncated, the ellipsis "..." could appear on the
+    wrong side for right-to-left scripts.
+  Moreover, when the text contains line breaks it now appears all on one line.
+  Note this only affects the labels in the plots; the rest of the report did not
+  have these problems.
+  :pr:`1097` by :user:`Jérôme Dockès <jeromedockes>`.
+
+* In the TableReport it is now possible, before clicking any of the cells, to
+  reach the dataframe sample table and activate a cell with tab key navigation.
+  :pr:`1101` by :user:`Jérôme Dockès <jeromedockes>`.
+
+* The "Column name" column of the "summary statistics" table in the TableReport
+  is now always visible when scrolling the table. :pr:`1102` by :user:`Jérôme
+  Dockès <jeromedockes>`.
+
+Release 0.3.1
+=============
+>>>>>>> main
 
 Minor changes
 -------------
@@ -48,6 +87,10 @@ Minor changes
   ``pandas.info()`` ) in a table. It can be sorted by each column.
   :pr:`1056` and :pr:`1068` by :user:`Jérôme Dockès <jeromedockes>`.
 
+* The credit fraud dataset is now available with the
+  :func:`fetch_credit_fraud function`.
+  :pr:`1053` by :user:`Vincent Maladiere <Vincent-Maladiere>`.
+
 * Added zero padding for column names in :class:`MinHashEncoder` to improve column ordering consistency.
   :pr:`1069` by :user:`Shreekant Nandiyawar <Shree7676>`.
 
@@ -61,6 +104,16 @@ Minor changes
 * It is possible to control the number of rows displayed by the TableReport in
   the "sample" tab panel by specifying ``n_rows``.
   :pr:`1083` by :user:`Jérôme Dockès <jeromedockes>`.
+
+* the `TableReport` used to raise an exception when the dataframe contained
+  unhashable types such as python lists. This has been fixed in :pr:`1087` by
+  :user:`Jérôme Dockès <jeromedockes>`.
+
+* Display's columns name with the HTML representation of the fitted TableVectorizer.
+  This has been fixed in :pr:`1093` by :user:`Shreekant Nandiyawar <Shree7676>`.
+
+* AggTarget will now work even when y is a Series and not raise any error.
+  This has been fixed in :pr:`1094` by :user:`Shreekant Nandiyawar <Shree7676>`.
 
 Release 0.3.0
 =============